+++ spec/rubocop/cli_spec.rb
@@ -14,11 +14,16 @@ describe RuboCop::CLI, :isolated_environment do
     $stdout = StringIO.new
     $stderr = StringIO.new
     RuboCop::ConfigLoader.debug = false
+
+    # OPTIMIZE: Makes these specs faster. Work directory (the parent of
+    # .rubocop_cache) is removed afterwards anyway.
+    RuboCop::ResultCache.inhibit_cleanup = true
   end
 
   after(:each) do
     $stdout = STDOUT
     $stderr = STDERR
+    RuboCop::ResultCache.inhibit_cleanup = false
   end
 
   def abs(path)